Turn on C++20. (NOTE: Manual change to wx/string.h is needed to get it to compile. Fixed in 3.1.6 which isn't in mainline Ubuntu anywhere).
Clean up the angle_functions, so they are more consistent in the use of float vs double (or in some cases making sure a floating point number is actually used when needed!)

+++ src/hssh/local_topological/area_detection/gateways/classifier_based_generator.cpp
@@ -285,7 +285,7 @@ struct ClassifierBasedGenerator::Impl
         //         }
         //     }
 
-        double skeletonNormal = 0.0;
+        float skeletonNormal = 0.0f;
 
         // Use regionCells_ to get the appropriate possibilities
         auto bestSources = skeleton.getSourceCells(maxCell.x, maxCell.y);
